Refactored app.js and modularized to server-methods.js
